Make HTTP2ConnectBufferingHandler public temporarily
